Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

It's mandatory to send some user-agent and profile in the MMS HTTP headers, but they appear not to actually matter.

The inheritence issue is that we can't change just one or two items in downstream version of mms_config.xml, the file has to be copied wholesale and modified separately in each device repo.  That's fine, but it's unfortunate for maintenance purposes.  It's like this because the _this_ XML file is considered an atomic (?) resource that's parsed at runtime by the Mms app, which is pretty silly but would require some effort to change.

In any event, let's submit for now since this works, so that SMS splitting and MMS is repaired for now and remains consistent with other d2 devices.  If you and Austen feel strongly about fixing the user agent in a d2vzw-specific mms_config.xml, feel free to add it, I wouldn't be opposed.

Patch-set: 1